+++ selenium/src/main/java/org/musetest/selenium/conditions/ElementExistsConditionStringExpressionSupport.java
@@ -1,5 +1,6 @@
 package org.musetest.selenium.conditions;
 
+import org.musetest.builtins.value.*;
 import org.musetest.core.*;
 import org.musetest.core.values.*;
 
@@ -9,7 +10,7 @@ import java.util.*;
  * @author Christopher L Merrill (see LICENSE.txt for license details)
  */
 @SuppressWarnings("unused")  // used via reflection
-public class ElementExistsConditionStringExpressionSupport extends BaseElementConditionStringExpressionSupport
+public class ElementExistsConditionStringExpressionSupport extends BaseArgumentedValueSourceStringSupport
     {
     @Override
     public String getName()
@@ -27,6 +28,12 @@ public class ElementExistsConditionStringExpressionSupport extends BaseElementCo
     protected String getTypeId()
         {
         return ElementExistsCondition.TYPE_ID;
+        }
+
+    @Override
+    protected boolean storeSingleArgumentAsSingleSubsource()
+        {
+        return true;
         }
     }